Ember.merge -> import merge
