+++ packages_es6/ember-views/lib/views/states/pre_render.js
@@ -1,8 +1,6 @@
 import _default from "ember-views/views/states/default";
-
-import Ember from "ember-metal/core"; // Ember.merge, Ember.create, Ember.$, Ember.assert
 import {create} from "ember-metal/platform";
-var merge = Ember.merge;
+import merge from "ember-metal/merge";
 
 /**
 @module ember
